+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 
 - Previously uncaught issue where sorting cards in Quest list was broken.
 - Issue where streaks wouldn't continue if points were greater than total or 1 >= `complete` > 0.
+- Set max requirements for quest type `q-w-s` to 6.
 
 ## [v.0.2.5]